requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
blessed==1.20.0
botocore==1.29.131
certifi==2023.5.7
chalice==1.28.0
charset-normalizer==3.1.0
click==8.1.3
frozenlist==1.3.3
future==0.18.3
idna==3.4
inquirer==2.10.1
jmespath==1.0.1
line-bot-sdk==2.4.2
lxml==4.9.2
multidict==6.0.4
python-dateutil==2.8.2
python-editor==1.0.4
PyYAML==6.0
readchar==4.0.5
requests==2.28.2
six==1.16.0
soupsieve==2.4.1
typing_extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
yarl==1.9.2
